requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asn1crypto==1.5.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.6.0
cryptography==42.0.2
fastapi==0.109.2
frozenlist==1.4.1
google-auth==2.27.0
h11==0.14.0
idna==3.6
multidict==6.0.5
pg8000==1.30.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
yarl==1.9.4
